djcelery.snapshot.Camera: Store events in the database using celeryev.
Using the snapshot feature available in celery master branch.
The resulting worker and task state is stored in the models
djcelery.models.TaskState and djcelery.models.WorkerState.

To start dumping events to your database you have to run celeryev
with the following arguments:

    python manage.py celeryev -c djcelery.snapshot.Camera --freq=2.0

This will update the database with the current cluster state every
2 seconds.